Support Stdout and Exception Details when Test Setup Fails
Adds support to buck test to capture stdout / report exceptions
correctly when setup of test fails - if @BeforeTest or @BeforeMethod
fails, any child test will also be marked as failing with the failure
exception and output from the setup method. This aligns with the
expected behavior of testng, where if setup fails, any child tests will
be skipped.

closes #3