[Buck] Log out Thread Name when Running ShellStep (#4)
First cut at helping with diagnosing issues in jenkins when the hang detection kicks in for buck. Currently we get the thread dump, but we cannot tell what actual ShellStep is associated with the thread - by logging the thread name, we should be able to match a waiting thread to the last command that was executed in that thread, which should let us figure out which test is hanging.